+++ packages/angular-workspace/nimble-angular/src/thirdparty/directives/router_link.ts
@@ -1,7 +1,8 @@
 /**
  * [Nimble]
- * Copied from https://github.com/angular/angular/blob/17.3.11/packages/router/src/directives/router_link.ts
+ * Copied from https://github.com/angular/angular/blob/18.2.13/packages/router/src/directives/router_link.ts
  * with the following modifications:
+ * - Copy `isUrlTree` function from url_tree.ts into this file instead of importing
  * - Hardcode `isAnchorElement` to `true` so that the directive will correctly set the `href` on elements within nimble that represent anchors
  * - Make `href` a `@HostBindinding` to avoid using Angular's private sanitization APIs because `href` bindings automatically are sanitized by
  *   Angular (see https://angular.io/guide/security#sanitization-and-security-contexts). Implementations leveraging RouterLink should have a test
@@ -15,7 +16,7 @@
  * Copyright Google LLC All Rights Reserved.
  *
  * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
+ * found in the LICENSE file at https://angular.dev/license
  */
 
 import {LocationStrategy} from '@angular/common';
@@ -31,13 +32,24 @@ import {
   type OnChanges,
   type OnDestroy,
   Renderer2,
+  ɵRuntimeError as RuntimeError,
   type SimpleChanges,
   ɵɵsanitizeUrlOrResourceUrl,
 } from '@angular/core';
 import {Subject, Subscription} from 'rxjs';
 
-import { type Event, type Params, type QueryParamsHandling, ActivatedRoute, Router, NavigationEnd, UrlTree } from '@angular/router';
+import {type Event, NavigationEnd} from '@angular/router';
+import type {QueryParamsHandling} from '@angular/router';
+import {Router} from '@angular/router';
+import {ActivatedRoute} from '@angular/router';
+import type {Params} from '@angular/router';
+import {UrlTree} from '@angular/router';
+// import {RuntimeErrorCode} from '../errors';
 
+// [Nimble] Copied from https://github.com/angular/angular/blob/18.2.13/packages/router/src/url_tree.ts
+function isUrlTree(v: any): v is UrlTree {
+  return v instanceof UrlTree;
+}
 
 /**
  * @description
@@ -105,6 +117,9 @@ import { type Event, type Params, type QueryParamsHandling, ActivatedRoute, Rout
  *   link to user component
  * </a>
  * ```
+ *
+ * `queryParams`, `fragment`, `queryParamsHandling`, `preserveFragment`, and `relativeTo`
+ * cannot be used when the `routerLink` input is a `UrlTree`.
  *
  * See {@link UrlCreationOptions#queryParamsHandling}.
  *
@@ -204,8 +219,6 @@ export class RouterLink implements OnChanges, OnDestroy {
    */
   @Input() relativeTo?: ActivatedRoute | null;
 
-  private commands: any[] | null = null;
-
   /** Whether a host element is an `<a>` tag. */
   private isAnchorElement: boolean;
 
@@ -275,7 +288,24 @@ export class RouterLink implements OnChanges, OnDestroy {
   }
 
   /** @nodoc */
-  ngOnChanges(changes: SimpleChanges) {
+  // TODO(atscott): Remove changes parameter in major version as a breaking change.
+  ngOnChanges(changes?: SimpleChanges) {
+    /* [Nimble] Comment out extra error handling that uses ngDevMode and RuntimeErrorCode
+    if (
+      ngDevMode &&
+      isUrlTree(this.routerLinkInput) &&
+      (this.fragment !== undefined ||
+        this.queryParams ||
+        this.queryParamsHandling ||
+        this.preserveFragment ||
+        this.relativeTo)
+    ) {
+      throw new RuntimeError(
+        RuntimeErrorCode.INVALID_ROUTER_LINK_INPUTS,
+        'Cannot configure queryParams or fragment when using a UrlTree as the routerLink input value.',
+      );
+    }
+    */
     if (this.isAnchorElement) {
       this.updateHref();
     }
@@ -284,21 +314,31 @@ export class RouterLink implements OnChanges, OnDestroy {
     this.onChanges.next(this);
   }
 
+  private routerLinkInput: any[] | UrlTree | null = null;
+
   /**
-   * Commands to pass to {@link Router#createUrlTree}.
+   * Commands to pass to {@link Router#createUrlTree} or a `UrlTree`.
    *   - **array**: commands to pass to {@link Router#createUrlTree}.
    *   - **string**: shorthand for array of commands with just the string, i.e. `['/route']`
+   *   - **UrlTree**: a `UrlTree` for this link rather than creating one from the commands
+   *     and other inputs that correspond to properties of `UrlCreationOptions`.
    *   - **null|undefined**: effectively disables the `routerLink`
    * @see {@link Router#createUrlTree}
    */
   @Input()
-  set routerLink(commands: any[] | string | null | undefined) {
-    if (commands != null) {
-      this.commands = Array.isArray(commands) ? commands : [commands];
-      this.setTabIndexIfNotOnNativeEl('0');
-    } else {
-      this.commands = null;
+  set routerLink(commandsOrUrlTree: any[] | string | UrlTree | null | undefined) {
+    if (commandsOrUrlTree == null) {
+      this.routerLinkInput = null;
       this.setTabIndexIfNotOnNativeEl(null);
+    } else {
+      if (isUrlTree(commandsOrUrlTree)) {
+        this.routerLinkInput = commandsOrUrlTree;
+      } else {
+        this.routerLinkInput = Array.isArray(commandsOrUrlTree)
+          ? commandsOrUrlTree
+          : [commandsOrUrlTree];
+      }
+      this.setTabIndexIfNotOnNativeEl('0');
     }
   }
 
@@ -364,23 +404,23 @@ export class RouterLink implements OnChanges, OnDestroy {
      * to be a `@HostBinding`. This avoids the need of calling into private Angular sanitization code.
      */
     // const sanitizedValue =
-    // this.href === null
-    //    ? null
-    //    : // This class represents a directive that can be added to both `<a>` elements,
-    //      // as well as other elements. As a result, we can't define security context at
-    //      // compile time. So the security context is deferred to runtime.
-    //      // The `ɵɵsanitizeUrlOrResourceUrl` selects the necessary sanitizer function
-    //      // based on the tag and property names. The logic mimics the one from
-    //      // `packages/compiler/src/schema/dom_security_schema.ts`, which is used at compile time.
-    //      //
-    //      // Note: we should investigate whether we can switch to using `@HostBinding('attr.href')`
-    //      // instead of applying a value via a renderer, after a final merge of the
-    //      // `RouterLinkWithHref` directive.
-    //      ɵɵsanitizeUrlOrResourceUrl(
-    //        this.href,
-    //        this.el.nativeElement.tagName.toLowerCase(),
-    //        'href',
-    //      );
+    //   this.href === null
+    //     ? null
+    //     : // This class represents a directive that can be added to both `<a>` elements,
+    //       // as well as other elements. As a result, we can't define security context at
+    //       // compile time. So the security context is deferred to runtime.
+    //       // The `ɵɵsanitizeUrlOrResourceUrl` selects the necessary sanitizer function
+    //       // based on the tag and property names. The logic mimics the one from
+    //       // `packages/compiler/src/schema/dom_security_schema.ts`, which is used at compile time.
+    //       //
+    //       // Note: we should investigate whether we can switch to using `@HostBinding('attr.href')`
+    //       // instead of applying a value via a renderer, after a final merge of the
+    //       // `RouterLinkWithHref` directive.
+    //       ɵɵsanitizeUrlOrResourceUrl(
+    //         this.href,
+    //         this.el.nativeElement.tagName.toLowerCase(),
+    //         'href',
+    //       );
     // this.applyAttributeValue('href', sanitizedValue);
   }
 
@@ -395,10 +435,12 @@ export class RouterLink implements OnChanges, OnDestroy {
   }
 
   get urlTree(): UrlTree | null {
-    if (this.commands === null) {
+    if (this.routerLinkInput === null) {
       return null;
+    } else if (isUrlTree(this.routerLinkInput)) {
+      return this.routerLinkInput;
     }
-    return this.router.createUrlTree(this.commands, {
+    return this.router.createUrlTree(this.routerLinkInput, {
       // If the `relativeTo` input is not defined, we want to use `this.route` by default.
       // Otherwise, we should use the value provided by the user in the input.
       relativeTo: this.relativeTo !== undefined ? this.relativeTo : this.route,