+++ packages/angular-workspace/nimble-angular/src/thirdparty/testing/matchers.ts
@@ -1,6 +1,6 @@
 /**
  * [Nimble]
- * Copied from https://github.com/angular/angular/blob/17.3.11/packages/platform-browser/testing/src/matchers.ts
+ * Copied from https://github.com/angular/angular/blob/18.2.13/packages/platform-browser/testing/src/matchers.ts
  * with the following modifications:
  * - Update imports
  * - Comment out everything other than what is needed to use `toHaveText` matcher because `toHaveText` is the only matcher required by the copied
@@ -14,9 +14,13 @@
  * Copyright Google LLC All Rights Reserved.
  *
  * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
+ * found in the LICENSE file at https://angular.dev/license
  */
 
+import {ɵgetDOM as getDOM} from '@angular/common';
+import {Type} from '@angular/core';
+import {ComponentFixture} from '@angular/core/testing';
+import {By} from '@angular/platform-browser';
 
 // [Nimble] Update imports
 import {childNodesAsList} from './browser_util';
@@ -55,7 +59,7 @@ export interface NgMatchers<T = any> extends jasmine.Matchers<T> {
   //  *
   //  * {@example testing/ts/matchers.ts region='toHaveCssStyle'}
   //  */
-  // toHaveCssStyle(expected: {[k: string]: string}|string): boolean;
+  // toHaveCssStyle(expected: {[k: string]: string} | string): boolean;
 
   // /**
   //  * Expect a class to implement the interface of the given class.
@@ -88,48 +92,49 @@ export interface NgMatchers<T = any> extends jasmine.Matchers<T> {
 const _expect: <T = any>(actual: T) => NgMatchers<T> = expect as any;
 export {_expect as expect};
 
-beforeEach(function() {
+beforeEach(function () {
   jasmine.addMatchers({
-    toHaveText: function() {
+    toHaveText: function () {
       return {
-        compare: function(actual: any, expectedText: string) {
+        compare: function (actual: any, expectedText: string) {
           const actualText = elementText(actual);
           return {
             pass: actualText == expectedText,
             get message() {
               return 'Expected ' + actualText + ' to be equal to ' + expectedText;
-            }
+            },
           };
-        }
+        },
       };
     },
 
     /* [Nimble] Comment out matchers that are not needed
-    toHaveCssClass: function() {
+    toHaveCssClass: function () {
       return {compare: buildError(false), negativeCompare: buildError(true)};
 
       function buildError(isNot: boolean) {
-        return function(actual: any, className: string) {
+        return function (actual: any, className: string) {
           return {
             pass: hasClass(actual, className) == !isNot,
             get message() {
               return `Expected ${actual.outerHTML} ${
-                  isNot ? 'not ' : ''}to contain the CSS class "${className}"`;
-            }
+                isNot ? 'not ' : ''
+              }to contain the CSS class "${className}"`;
+            },
           };
         };
       }
     },
 
-    toHaveCssStyle: function() {
+    toHaveCssStyle: function () {
       return {
-        compare: function(actual: any, styles: {[k: string]: string}|string) {
+        compare: function (actual: any, styles: {[k: string]: string} | string) {
           let allPassed: boolean;
           if (typeof styles === 'string') {
             allPassed = hasStyle(actual, styles);
           } else {
             allPassed = Object.keys(styles).length !== 0;
-            Object.keys(styles).forEach(prop => {
+            Object.keys(styles).forEach((prop) => {
               allPassed = allPassed && hasStyle(actual, prop, styles[prop]);
             });
           }
@@ -139,17 +144,16 @@ beforeEach(function() {
             get message() {
               const expectedValueStr = typeof styles === 'string' ? styles : JSON.stringify(styles);
               return `Expected ${actual.outerHTML} ${!allPassed ? ' ' : 'not '}to contain the
-                      CSS ${typeof styles === 'string' ? 'property' : 'styles'} "${
-                  expectedValueStr}"`;
-            }
+                      CSS ${typeof styles === 'string' ? 'property' : 'styles'} "${expectedValueStr}"`;
+            },
           };
-        }
+        },
       };
     },
 
-    toImplement: function() {
+    toImplement: function () {
       return {
-        compare: function(actualObject: any, expectedInterface: any) {
+        compare: function (actualObject: any, expectedInterface: any) {
           const intProps = Object.keys(expectedInterface.prototype);
 
           const missedMethods: any[] = [];
@@ -160,17 +164,21 @@ beforeEach(function() {
           return {
             pass: missedMethods.length == 0,
             get message() {
-              return 'Expected ' + actualObject +
-                  ' to have the following methods: ' + missedMethods.join(', ');
-            }
+              return (
+                'Expected ' +
+                actualObject +
+                ' to have the following methods: ' +
+                missedMethods.join(', ')
+              );
+            },
           };
-        }
+        },
       };
     },
 
-    toContainComponent: function() {
+    toContainComponent: function () {
       return {
-        compare: function(actualFixture: any, expectedComponentType: Type<any>) {
+        compare: function (actualFixture: any, expectedComponentType: Type<any>) {
           const failOutput = arguments[2];
           const msgFn = (msg: string): string => [msg, failOutput].filter(Boolean).join(', ');
 
@@ -178,18 +186,19 @@ beforeEach(function() {
           if (!(actualFixture instanceof ComponentFixture)) {
             return {
               pass: false,
-              message: msgFn(`Expected actual to be of type \'ComponentFixture\' [actual=${
-                  actualFixture.constructor.name}]`)
+              message: msgFn(
+                `Expected actual to be of type \'ComponentFixture\' [actual=${actualFixture.constructor.name}]`,
+              ),
             };
           }
 
           const found = !!actualFixture.debugElement.query(By.directive(expectedComponentType));
-          return found ?
-              {pass: true} :
-              {pass: false, message: msgFn(`Expected ${expectedComponentType.name} to show`)};
-        }
+          return found
+            ? {pass: true}
+            : {pass: false, message: msgFn(`Expected ${expectedComponentType.name} to show`)};
+        },
       };
-    }
+    },
     */
   });
 });