+++ packages/nimble-components/src/spinner/styles.ts
@@ -77,12 +77,14 @@ export const styles = css`
         }
 
         div.bit1 {
-            animation-name: ni-private-spinner-no-motion-opacity-keyframes,
+            animation-name:
+                ni-private-spinner-no-motion-opacity-keyframes,
                 ni-private-spinner-no-motion-transform-keyframes-1;
         }
 
         div.bit2 {
-            animation-name: ni-private-spinner-no-motion-opacity-keyframes,
+            animation-name:
+                ni-private-spinner-no-motion-opacity-keyframes,
                 ni-private-spinner-no-motion-transform-keyframes-2;
         }
     }