+++ packages/nimble-components/src/table-column/menu-button/types.ts
@@ -1,7 +1,7 @@
 import type { MenuButtonToggleEventDetail } from '../../menu-button/types';
 
 /** @internal */
-export const cellViewMenuSlotName = 'menu-button-menu' as const;
+export const cellViewMenuSlotName = 'menu-button-menu';
 
 /**
  * The type of the detail associated with the `menu-button-column-beforetoggle` and `menu-button-column-toggle`